Fix a broken test
