+++ src/bundles/tests/test_views.py
@@ -32,33 +32,40 @@ def test_anonymous_user_cannot_bookmark_aids(client, aid):
     assert res.status_code == 405  # method not allowed
 
 
-def test_bundle_modal_shows_selected_bundles(client, aid):
+def test_bundle_modal_shows_selected_bundles(
+        live_server, browser, client, aid):
     """Authenticated users cannot login again, duh!"""
 
     user = aid.author
-    bundles = [
+    [
         BundleFactory(name='Bundle_1', owner=user, aids=[aid]),
         BundleFactory(name='Bundle_2', owner=user, aids=[aid]),
         BundleFactory(name='Bundle_3', owner=user, aids=[])
     ]
 
+    # Browser login
     client.force_login(user)
-    aid_url = aid.get_absolute_url()
-    res = client.get(aid_url)
-    content = res.content.decode()
+    cookie = client.cookies['sessionid']
+    aid_url = live_server + aid.get_absolute_url()
+    browser.get(aid_url)
+    browser.add_cookie({'name': 'sessionid', 'value': cookie.value,
+                        'secure': False, 'path': '/'})
+    browser.refresh()
+
+    modal = browser.find_element_by_id('bookmark-modal')
+    content = modal.get_attribute('innerHTML')
     assert 'Vous devez être identifé·e pour utiliser ' \
            'les listes d\'aides' not in content
-
     assert '<form id="bookmark-form"' in content
-    assert '<input type="checkbox" name="bundles" value="{}" ' \
-           'id="id_bundles_0" checked>'.format(bundles[0].id) in content
-    assert '<label for="id_bundles_0">Bundle_1</label>' in content
-    assert '<input type="checkbox" name="bundles" value="{}" ' \
-           'id="id_bundles_1" checked>'.format(bundles[1].id) in content
-    assert '<label for="id_bundles_1">Bundle_2</label>' in content
-    assert '<input type="checkbox" name="bundles" value="{}" ' \
-           'id="id_bundles_2">'.format(bundles[2].id) in content
-    assert '<label for="id_bundles_2">Bundle_3</label>' in content
+
+    input0 = browser.find_element_by_id('id_bundles_0')
+    assert input0.is_selected()
+
+    input1 = browser.find_element_by_id('id_bundles_1')
+    assert input1.is_selected()
+
+    input2 = browser.find_element_by_id('id_bundles_2')
+    assert not input2.is_selected()
 
 
 def test_bundle_selection_is_effective(client, aid):
